# docs/swagger.yaml
basePath: /
definitions:
  http.APIResponse:
    properties:
      status:
        type: string
    type: object
...
  http.approveRequest:
    properties:
      action:
        description: defaults to approve
        type: string
      id:
        type: string
      identifier:
        type: string
      voter:
        type: string
    type: object
...
  http.azureWebhook:
    properties:
      request:
        properties:
          host:
            type: string
        type: object
      target:
        properties:
          digest:
            type: string
          repository:
            type: string
          tag:
            type: string
        type: object
    type: object
...
  http.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  http.loginResponse:
    properties:
      token:
        type: string
    type: object
...
  http.resourceApprovalsUpdateRequest:
    properties:
      identifier:
        type: string
      provider:
        type: string
      votesRequired:
        type: integer
    type: object
...
  types.Approval:
    properties:
      archived:
        description: Archived is set to true once approval is finally approved/rejected
        type: boolean
      createdAt:
        description: When this approval was created
        type: string
      currentVersion:
        type: string
      deadline:
        description: Deadline for this request
        type: string
      digest:
        description: |-
          Digest is used to verify that images are the ones that got the approvals.
          If digest doesn't match for the image, votes are reset.
        type: string
      event:
        allOf:
        - $ref: '#/definitions/types.Event'
        description: Event that triggered evaluation
      id:
        type: string
      identifier:
        description: |-
          Identifier is used to inform user about specific
          Helm release or k8s deployment
          ie: k8s <namespace>/<deployment name>
              helm: <namespace>/<release name>
        type: string
      message:
        type: string
      newVersion:
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/types.ProviderType'
        description: Provider name - Kubernetes/Helm
      rejected:
        description: |-
          Explicitly rejected approval
          can be set directly by user
          so even if deadline is not reached approval
          could be turned down
        type: boolean
      updatedAt:
        description: WHen this approval was updated
        type: string
      voters:
        allOf:
        - $ref: '#/definitions/types.JSONB'
        description: |-
          Voters is a list of voter
          IDs for audit
      votesReceived:
        type: integer
      votesRequired:
        description: |-
          Requirements for the update such as number of votes
          and deadline
        type: integer
    type: object
...
  types.Event:
    properties:
      createdAt:
        type: string
      repository:
        $ref: '#/definitions/types.Repository'
      triggerName:
        description: optional field to identify trigger
        type: string
    type: object
  types.JSONB:
    additionalProperties: true
    type: object
  types.ProviderType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - ProviderTypeUnknown
    - ProviderTypeKubernetes
    - ProviderTypeHelm
  types.Repository:
    properties:
      digest:
        description: optional digest field
        type: string
      host:
        type: string
      name:
        type: string
      tag:
        type: string
    type: object
...
host: localhost:9300
info:
  contact:
    email: support@keel.sh
    name: Keel Support
    url: https://keel.sh
  description: Automated Kubernetes deployment updates. Keel is a tool for automating
    Kubernetes deployment updates.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://keel.sh/terms/
  title: Keel API
  version: "1.0"
paths:
...
  /v1/approvals:
    get:
      description: Returns a list of all approvals (both active and archived)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Approval'
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: List approvals
      tags:
      - approvals
    post:
      consumes:
      - application/json
      description: Performs an action on an approval (approve, reject, delete, or
        archive)
      parameters:
      - description: Approval action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.approveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Approval'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Approval not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Approve, reject, delete, or archive an approval
      tags:
      - approvals
    put:
      consumes:
      - application/json
      description: Sets or removes approval requirements for a resource
      parameters:
      - description: Approval update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.resourceApprovalsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.APIResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Resource not found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update approval requirements
      tags:
      - approvals
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with username and password
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.loginResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Username or password incorrect
          schema:
            type: string
      summary: Login user
      tags:
      - auth
  /v1/auth/logout:
    get:
      description: Logs out the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - auth
    post:
      description: Logs out the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - auth
  /v1/auth/refresh:
    get:
      description: Generates a new authentication token for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.loginResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Refresh authentication token
      tags:
      - auth
...
  /v1/webhooks/azure:
    post:
      consumes:
      - application/json
      description: Receives and processes Azure Container Registry webhook notifications
        for image push events
      parameters:
      - description: Azure webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/http.azureWebhook'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Trigger Azure Container Registry webhook
      tags:
      - webhooks
...
swagger: "2.0"